+++ framework/model/src/main/java/is/codion/framework/model/AbstractEntityEditModel.java
@@ -440,6 +440,11 @@ public abstract class AbstractEntityEditModel implements EntityEditModel {
 	}
 
 	@Override
+	public final EventObserver<Attribute<?>> valueEvent() {
+		return events.valueChange.observer();
+	}
+
+	@Override
 	public final EventObserver<Entity> entityEvent() {
 		return events.entity.observer();
 	}